+++ Tutorial01/Tutorial01/SkeletalMeshComponent.h
@@ -29,7 +29,7 @@ public:
 
 	AnimClipInstance* _CurrentAnim;
 public:
-	void PlayAnim(AnimationClip* InClip, int InNumPlay = 0, int InRate = 1.f);
+	void PlayAnim(AnimationClip* InClip, int InNumPlay = 0, float InRate = 1.f);
 	void Tick(float DeltaSeconds);
 	void UpdateBoneMatrices();
